nyayai-backend/requirements.txt
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
pandas>=2.0
scikit-learn>=1.3
fairlearn>=0.9
google-generativeai>=0.5
werkzeug>=3.0
scipy>=1.11
numpy>=1.24
